# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "KPIs API Documentation",
        "description": "KPIs API Documentation",
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/aggregators/list": {
            "get": {
                "tags": [
                    "aggregators"
                ],
                "summary": "List Aggregators",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "query",
                        "description": "Localization paremeter. Choose from available languages (en, el).",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "en",
                            "el"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with all available aggregators."
                    }
                }
            }
        },
...
    },
...
}